karton-confext-example/requirements.txt
# required
karton-core>=4.2.0,<5.0.0
yara-python
mwdblib

# custom